==> src/notebooks/01_json_scraping.ipynb <==
import json

file_path = "../../data/json_handling/spotify_playlist.json"

with open(file_path, 'r') as json_file:
    data = json.load(json_file)

# the name of the 31st track
print(data['contents']['items'][30]['name'])

# total playcount
total_playCount = 0
for track in data['contents']['items']:
    total_playCount += track['playCount']

print(total_playCount)

# minimum playcount
target_index = 0
min_playCount = data['contents']['items'][0]['playCount']
for i in range(len(data['contents']['items'])):
    if data['contents']['items'][i]['playCount'] < min_playCount:
        min_playCount = data['contents']['items'][i]['playCount']
        target_index = i

song_name = data['contents']['items'][target_index]['name']

print(f'The lowest playcount is {song_name}')
