==> src/notebooks/06_trips_data_mapping.ipynb <==
from datetime import datetime
from dateutil.relativedelta import relativedelta
import os

import pandas as pd
pd.set_option('display.max_columns', 30)
import requests

current_datetime = datetime.now() - relativedelta(month=2)

formatted_datetime = current_datetime.strftime('%Y-%m-%d')

url = f"https://data.cityofchicago.org/resource/ajtu-isnz.json?$where=trip_start_timestamp >= '{formatted_datetime}T00:00:00' AND trip_start_timestamp <= '{formatted_datetime}T23:59:59'&$limit=30000"

headers = {"X-App-Token": os.environ.get("CHICAGO_API_TOKEN")}

response = requests.get(url)

data = response.json()

taxi_trips = pd.DataFrame(data)

taxi_trips.head()

taxi_trips.info()

taxi_trips.describe()

taxi_trips[taxi_trips['fare'].isna()].sample(5)

#### Transformation: deal with NaN values

taxi_trips.drop(['pickup_census_tract', 'dropoff_census_tract'], axis=1, inplace=True)

taxi_trips.drop(['pickup_centroid_location', 'dropoff_centroid_location'], axis=1, inplace=True)

taxi_trips.dropna(inplace=True)
taxi_trips.info()

#### Transformation: renaming

taxi_trips.rename(columns={'pickup_community_area': 'pickup_community_area_id', 
                           'dropoff_community_area': 'dropoff_community_area_id'}, inplace=True)

#### Transformation: create helper column for weather

taxi_trips['trip_start_timestamp'] = pd.to_datetime(taxi_trips['trip_start_timestamp'])

taxi_trips['datetime_for_weather'] = taxi_trips['trip_start_timestamp'].dt.floor('h')

#### Check joining the trips and the weather data

# Get weather data, and create a dataframe out of it

current_datetime = datetime.now() - relativedelta(month=2)
formatted_datetime = current_datetime.strftime('%Y-%m-%d')

url = "https://archive-api.open-meteo.com/v1/era5"

params = {
    'latitude': 41.85,
    'longitude': -87.65,
    'start_date': formatted_datetime,
    'end_date': formatted_datetime,
    'hourly': 'temperature_2m,wind_speed_10m,rain,precipitation'
}

response = requests.get(url, params=params)
weather_data = response.json()

weather_data_filtered = {
    'datetime': weather_data['hourly']['time'],
    'temperature': weather_data['hourly']['temperature_2m'],
    'wind_speed': weather_data['hourly']['wind_speed_10m'],
    'rain': weather_data['hourly']['rain'],
    'precipitation': weather_data['hourly']['precipitation']
}

weather_df = pd.DataFrame(weather_data_filtered)
weather_df['datetime'] = pd.to_datetime(weather_df['datetime'] )

weather_df.head()

taxi_trips_with_weather = taxi_trips.merge(weather_df, left_on='datetime_for_weather', right_on='datetime')

taxi_trips_with_weather.head()

#### Transformation: data type conversion

taxi_trips.info()

taxi_trips.head()

data_types = {
    'trip_end_timestamp': 'datetime64[ns]',
    'trip_seconds': 'int32',
    'trip_miles': 'float',
    'pickup_community_area_id': 'int8',
    'dropoff_community_area_id': 'int8',
    'fare': 'float',
    'tips': 'float',
    'tolls': 'float',
    'extras': 'float',
    'trip_total': 'float'
}

taxi_trips = taxi_trips.astype(data_types)

taxi_trips.info()

#### Sanity check

taxi_trips.describe()

taxi_trips[taxi_trips['trip_end_timestamp'] == taxi_trips['trip_end_timestamp'].max()]

taxi_trips[taxi_trips['trip_seconds'] == taxi_trips['trip_seconds'].max()]

taxi_trips[taxi_trips['fare'] == taxi_trips['fare'].max()]

taxi_trips.nlargest(5, 'fare')

### Create mapping tables

payment_type_map_table = taxi_trips['payment_type'].drop_duplicates().reset_index(drop=True)

payment_type_map_table = pd.DataFrame(
    {
        'payment_type_id': range(1, len(payment_type_map_table) + 1),
        'payment_type': payment_type_map_table
    }
)
payment_type_map_table

company_map_table = taxi_trips['company'].drop_duplicates().reset_index(drop=True)

company_map_table = pd.DataFrame(
    {
        'company_id': range(1, len(company_map_table) + 1),
        'company': company_map_table
    }
)
company_map_table

payment_type_map_table.to_csv('payment_type_map_table.csv', index=False)
company_map_table.to_csv('company_map_table.csv', index=False)

### Checking the modeling

#### Payment type

payment_type_map_table = taxi_trips['payment_type'].drop_duplicates().reset_index(drop=True)

payment_type_map_table = pd.DataFrame(
    {
        'payment_type_id': range(1, len(payment_type_map_table) + 1),
        'payment_type': payment_type_map_table
    }
)
payment_type_map_table

new_payment_type_data = [
    {'payment_type': 'Credit Card'},
    {'payment_type': 'X'},
    {'payment_type': 'Y'}
]

new_payment_type_mapping = pd.DataFrame(new_payment_type_data)
new_payment_type_mapping

payment_type_max_id = payment_type_map_table['payment_type_id'].max()

range(payment_type_max_id + 1, payment_type_max_id + len(new_payment_type_mapping) + 1)

new_payment_type_mapping['payment_type_id'] = range(payment_type_max_id + 1, payment_type_max_id + len(new_payment_type_mapping) + 1)
new_payment_type_mapping

new_payment_types = new_payment_type_mapping[~new_payment_type_mapping['payment_type'].isin(payment_type_map_table['payment_type'])]
new_payment_types

if not new_payment_types.empty:
    payment_type_map_table = pd.concat([payment_type_map_table, new_payment_types], ignore_index=True)

payment_type_map_table

#### Company

company_map_table = taxi_trips['company'].drop_duplicates().reset_index(drop=True)

company_map_table = pd.DataFrame(
    {
        'company_id': range(1, len(company_map_table) + 1),
        'company': company_map_table
    }
)
company_map_table.tail()

new_company_data = [
    {'company': '5167 - 71969 5167 Taxi Inc'},
    {'company': 'X'},
    {'company': 'Y'}
]

new_company_mapping = pd.DataFrame(new_company_data)
new_company_mapping

company_max_id = company_map_table['company_id'].max()

new_company_mapping['company_id'] = range(company_max_id + 1, company_max_id + len(new_company_mapping) + 1)
new_company_mapping

new_companies = new_company_mapping[~new_company_mapping['company'].isin(company_map_table['company'])]
new_companies

if not new_companies.empty:
    company_map_table = pd.concat([company_map_table, new_companies], ignore_index=True)
