==> src/notebooks/07_transform_load.ipynb <==
from datetime import datetime
from dateutil.relativedelta import relativedelta
import os

import pandas as pd
pd.set_option('display.max_columns', 30)
import requests

"""
1. Get the data from S3
2. Weather data transformation
3. Taxi trips data transformation
4. Update payment_type_master
5. Update company_master
6. Update taxi_trips with company and payment_type ids (replace the string values with ids)
7. Upload weather data to s3
8. Upload taxi data to s3
9. Upload the newest company_master and payment_type master

"""

### Taxi trips transformation codes

current_datetime = datetime.now() - relativedelta(month=2)

formatted_datetime = current_datetime.strftime('%Y-%m-%d')

url = f"https://data.cityofchicago.org/resource/ajtu-isnz.json?$where=trip_start_timestamp >= '{formatted_datetime}T00:00:00' AND trip_start_timestamp <= '{formatted_datetime}T23:59:59'&$limit=30000"

headers = {"X-App-Token": os.environ.get("CHICAGO_API_TOKEN")}

response = requests.get(url)

data = response.json()

taxi_trips = pd.DataFrame(data)

taxi_trips.drop(['pickup_census_tract', 'dropoff_census_tract'], axis=1, inplace=True)
taxi_trips.drop(['pickup_centroid_location', 'dropoff_centroid_location'], axis=1, inplace=True)

taxi_trips.dropna(inplace=True)

taxi_trips.rename(columns={'pickup_community_area': 'pickup_community_area_id', 
                           'dropoff_community_area': 'dropoff_community_area_id'}, inplace=True)

taxi_trips['datetime_for_weather'] = pd.to_datetime(taxi_trips['trip_start_timestamp']).dt.floor('h')

taxi_trips.head()

#### Taxi trips transformation function

def taxi_trips_transformations(taxi_trips: pd.DataFrame) -> pd.DataFrame:
    """Perform transformations with the taxi data.

    Parameters:
        taxi_trips (pd.DataFrame): The DataFrame holding the daily taxi trips.

    Returns:
        pd.DataFrame: The cleaned, transformed DataFrame holding the daily taxi trips.
    
    """

    if not isinstance(taxi_trips, pd.DataFrame):
        raise TypeError('taxi_trips is not a valid pandas DataFrame')

    taxi_trips.drop(['pickup_census_tract', 'dropoff_census_tract', 
                     'pickup_centroid_location', 'dropoff_centroid_location'], axis=1, inplace=True)

    taxi_trips.dropna(inplace=True)

    taxi_trips.rename(columns={'pickup_community_area': 'pickup_community_area_id', 
                            'dropoff_community_area': 'dropoff_community_area_id'}, inplace=True)

    taxi_trips['datetime_for_weather'] = pd.to_datetime(taxi_trips['trip_start_timestamp']).dt.floor('h')

    return taxi_trips

### Company update codes

company_map_table = taxi_trips['company'].drop_duplicates().reset_index(drop=True)

company_map_table = pd.DataFrame(
    {
        'company_id': range(1, len(company_map_table) + 1),
        'company': company_map_table
    }
)
company_map_table.tail()

new_company_data = [
    {'company': 'Setare Inc'},
    {'company': 'X'},
    {'company': 'Y'}
]

new_company_mapping = pd.DataFrame(new_company_data)
new_company_mapping

company_max_id = company_map_table['company_id'].max()


new_companies_list = [company for company in new_company_mapping['company'].values if company not in company_map_table['company'].values]
new_companies_list

new_companies_df = pd.DataFrame({
    'company_id': range(company_max_id + 1, company_max_id + len(new_companies_list) + 1),
    'company': new_companies_list

})
new_companies_df

updated_company_map_table = pd.concat([company_map_table, new_companies_df], ignore_index=True)
updated_company_map_table.tail()

def update_company_master(taxi_trips: pd.DataFrame, company_map_table: pd.DataFrame) -> pd.DataFrame:
    """Extend the company map table with new companies if there are new companies.

    Args:
        taxi_trips (pd.DataFrame): DataFrame holding the daily taxi trips.
        company_map_table (pd.DataFrame): DataFrame holding the company master data.

    Returns:
        pd.DataFrame: The updated company master data.
    """
    
    company_max_id = company_map_table['company_id'].max()

    new_companies_list = [company for company in taxi_trips['company'].values if company not in company_map_table['company'].values]

    new_companies_df = pd.DataFrame({
    'company_id': range(company_max_id + 1, company_max_id + len(new_companies_list) + 1),
    'company': new_companies_list

    })
    
    updated_company_map_table = pd.concat([company_map_table, new_companies_df], ignore_index=True)

    return updated_company_map_table

### Payment_type master codes

payment_type_map_table = taxi_trips['payment_type'].drop_duplicates().reset_index(drop=True)

payment_type_map_table = pd.DataFrame(
    {
        'payment_type_id': range(1, len(payment_type_map_table) + 1),
        'payment_type': payment_type_map_table
    }
)

taxi_trips_payment_type_only = pd.DataFrame({
    'payment_type_id': [1, 2, 3],
    'payment_type': ['Credit Card', 'X', 'Y']
})
taxi_trips_payment_type_only

def update_payment_type_master(taxi_trips: pd.DataFrame, payment_type_map_table: pd.DataFrame) -> pd.DataFrame:
    """Extend the payment type map table with new payment types if there are new payment types.

    Args:
        taxi_trips (pd.DataFrame): DataFrame holding the daily taxi trips.
        payment_type_map_table (pd.DataFrame): DataFrame holding the payment type master data.

    Returns:
        pd.DataFrame: The updated payment type master data.
    """
    
    payment_type_max_id = payment_type_map_table['company_id'].max()

    new_payment_types_list = [payment_type for payment_type in taxi_trips['payment_type'].values if payment_type not in payment_type_map_table['payment_type'].values]

    new_payment_types_df = pd.DataFrame({
    'payment_type_id': range(payment_type_max_id + 1, payment_type_max_id + len(new_payment_types_list) + 1),
    'payment_type': new_payment_types_list

    })
    
    updated_payment_type_map_table = pd.concat([payment_type_map_table, new_payment_types_df], ignore_index=True)

    return updated_payment_type_map_table

### Creating general update master table function

def update_master(taxi_trips: pd.DataFrame, map_table: pd.DataFrame, id_column: str, value_column: str) -> pd.DataFrame:
    """Extend the map table with new values if there are new values.

    Args:
        taxi_trips (pd.DataFrame): DataFrame holding the daily taxi trips.
        map_table (pd.DataFrame): DataFrame holding the master data.
        id_column (str): The id column of map table.
        value_column (str): Name of the column in map table containing the values.

    Returns:
        pd.DataFrame: The updated master data.
    """
    
    max_id = map_table[id_column].max()

    # new_values_list = [value for value in taxi_trips[value_column].values if value not in map_table[value_column].values]
    new_values_list = list(set(taxi_trips[value_column].values) - set(map_table[value_column].values))

    new_values_df = pd.DataFrame({
    id_column: range(max_id + 1, max_id + len(new_values_list) + 1),
    value_column: new_values_list

    })
    
    updated_map_table = pd.concat([map_table, new_values_df], ignore_index=True)

    return updated_map_table

### Update taxi trips with the most recent company and payment type map table function

def update_taxi_trips_wit_map_table(taxi_trips: pd.DataFrame, company_map_table: pd.DataFrame, payment_type_map_table: pd.DataFrame) -> pd.DataFrame:
    """Update the taxi trips DataFrame with company and payment type map table ids, and delete the string columns.

    Args:
        taxi_trips (pd.DataFrame): DataFrame holding the daily taxi trips.
        company_map_table (pd.DataFrame): The comapny map table.
        payment_type_map_table (pd.DataFrame): The payment type map table.

    Returns:
        pd.DataFrame: The taxi trips data with only payment type id and company id, with no company and payment type values.
    """
    taxi_trips_id = taxi_trips.merge(company_map_table, on='company')
    taxi_trips_id = taxi_trips_id.merge(payment_type_map_table, on='payment_type')

    taxi_trips_id.drop(['company', 'payment_type'], axis=1, inplace=True)

    return taxi_trips_id

### Weather transformation function


def transform_weather_data(weather_data: json) -> pd.DataFrame:
    """Make transformations on daily weather api response.

    Args:
        weather_data (json): The daily weather data from Open-Meteo API.

    Returns:
        pd.DataFrame: A DataFrame representation of the data.
    """

    weather_data_filtered = {
        'datetime': weather_data['hourly']['time'],
        'temperature': weather_data['hourly']['temperature_2m'],
        'wind_speed': weather_data['hourly']['wind_speed_10m'],
        'rain': weather_data['hourly']['rain'],
        'precipitation': weather_data['hourly']['precipitation']
    }

    weather_df = pd.DataFrame(weather_data_filtered)
    weather_df['datetime'] = pd.to_datetime(weather_df['datetime'] )

    return weather_df
